requirements.txt
numpy>=1.24.0
google-genai>=0.1.0
python-dotenv>=1.0.0
msgpack>=1.0.0
sentence-transformers>=2.2.0
